refactor: move Button css constant and apply to chip
